Move last E2E tests into DaemonTest and DistributionTest in assembly
