+++ test/assembly/AssemblyConstants.java
@@ -32,7 +32,7 @@ import java.util.concurrent.TimeoutException;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
-public class AssemblyE2EConstants {
+public class AssemblyConstants {
     public static final Path GRAKN_TARGET_DIRECTORY = Paths.get(".").toAbsolutePath();
     public static final Path ZIP_FULLPATH = Paths.get(GRAKN_TARGET_DIRECTORY.toString(), "grakn-core-all-mac.zip");
     public static final Path GRAKN_UNZIPPED_DIRECTORY = Paths.get(GRAKN_TARGET_DIRECTORY.toString(), "assembly-test", "grakn-core-all-mac");
